Enhance Documentation and Refactor Argument Management in Haplotag
- Add detailed documentation comments for various classes and methods in `ArgumentManager`, `Haplotag`, and `VcfParser`, improving code clarity and usability.
- Refactor `HaplotagProcess` to rename `taggingProcess` to `pipelineProcess`, aligning with updated terminology and enhancing consistency across the codebase.
- Update logging messages in `SomaticReadVerifier` and `SomaticVarCaller` to improve error reporting and clarity.
- Introduce new methods in `TumorPurityEstimator` and `SomaticVarCaller` for better handling of tumor purity estimation and somatic variant flags, enhancing functionality.
- Modify `Makefile` to include new source files related to the refactoring, ensuring proper compilation.
