+++ ArgumentManager.cpp
@@ -1,10 +1,19 @@
 #include "ArgumentManager.h"
 
-
+/**
+ * @brief Set the short options string for getopt
+ * @param opt The short options string (e.g., "s:b:o:t:q:p:r:")
+ */
 void ArgumentManager::setShortOption(const std::string& opt) {
     shortOpts = opt;
 }
 
+/**
+ * @brief Add a long option to the options list
+ * @param opt The option structure to add
+ * 
+ * Inserts the option before the terminator if it's not the terminator itself
+ */
 void ArgumentManager::addOption(const struct option& opt) {
     // If this is not the terminator, add it before the last element
     if (opt.name != NULL && !longOpts.empty()) {
@@ -14,7 +23,12 @@ void ArgumentManager::addOption(const struct option& opt) {
     }
 }
 
-
+/**
+ * @brief Validate that a required file exists and is accessible
+ * @param filePath Path to the file to validate
+ * @param fileDescription Description of the file for error messages
+ * @return true if file exists and is accessible, false otherwise
+ */
 bool ArgumentManager::validateRequiredFile(const std::string& filePath, const std::string& fileDescription) {
     if(filePath.empty()) {
         std::cerr << "[ERROR] " << programName  << ": missing " << fileDescription << ".\n";
@@ -30,6 +44,12 @@ bool ArgumentManager::validateRequiredFile(const std::string& filePath, const st
     return true;
 }
 
+/**
+ * @brief Validate that an optional file exists and is accessible (if specified)
+ * @param filePath Path to the file to validate
+ * @param fileDescription Description of the file for error messages
+ * @return true if file is not specified or exists and is accessible, false otherwise
+ */
 bool ArgumentManager::validateOptionalFile(const std::string& filePath, const std::string& fileDescription) {
     if(filePath.empty()) {
         // Optional file not specified, that's OK
@@ -44,6 +64,15 @@ bool ArgumentManager::validateOptionalFile(const std::string& filePath, const st
     return true;
 }
 
+/**
+ * @brief Parse command line arguments using getopt_long
+ * @param argc Number of command line arguments
+ * @param argv Array of command line argument strings
+ * 
+ * This function processes command line arguments, validates files and parameters,
+ * and handles help requests. It calls virtual functions that should be overridden
+ * by derived classes for specific parameter handling.
+ */
 void ArgumentManager::parseOptions(int argc, char** argv)
 {
 
@@ -101,8 +130,20 @@ void ArgumentManager::parseOptions(int argc, char** argv)
     } 
 }
 
+/**
+ * @brief Namespace containing file validation utilities
+ * 
+ * Provides standalone functions for file validation that can be used
+ * without an ArgumentManager instance
+ */
 namespace FileValidator{
-    // Validate if a required file exists
+    /**
+     * @brief Validate that a required file exists and is accessible
+     * @param filePath Path to the file to validate
+     * @param fileDescription Description of the file for error messages
+     * @param programName Name of the program for error messages
+     * @return true if file exists and is accessible, false otherwise
+     */
     bool validateRequiredFile(const std::string& filePath, const std::string& fileDescription, const std::string& programName) {
         if(filePath.empty()) {
             std::cerr << "[ERROR] " << programName  << ": missing " << fileDescription << ".\n";
@@ -118,7 +159,13 @@ namespace FileValidator{
         return true;
     }
 
-    // Validate if an optional file exists (if specified)
+    /**
+     * @brief Validate that an optional file exists and is accessible (if specified)
+     * @param filePath Path to the file to validate
+     * @param fileDescription Description of the file for error messages
+     * @param programName Name of the program for error messages
+     * @return true if file is not specified or exists and is accessible, false otherwise
+     */
     bool validateOptionalFile(const std::string& filePath, const std::string& fileDescription, const std::string& programName) {
         if(filePath.empty()) {
             // Optional file not specified, that's OK
