+++ HaplotagParsingBam.cpp
@@ -1,5 +1,20 @@
 #include "HaplotagParsingBam.h"
 
+/**
+ * @brief Constructor for BamFileRAII class
+ * @param BamFile Path to the input BAM file
+ * @param fastaFile Path to the reference FASTA file
+ * @param threadPool Thread pool for parallel processing
+ * @param version Program version string
+ * @param command Command line string
+ * @param resultPrefix Output file prefix
+ * @param outputFormat Output format (bam/cram)
+ * @param writeOutputBam Whether to write output BAM file
+ * 
+ * Initializes BAM file resources and sets up input/output streams.
+ * Handles BAM file opening, header reading, index loading, and thread pool setup.
+ * Supports both input and output BAM file management with proper error handling.
+ */
 BamFileRAII::BamFileRAII(
     const std::string& BamFile,
     const std::string& fastaFile,
@@ -41,7 +56,7 @@ writeOutputBam(writeOutputBam), isReleased(false), in(nullptr), out(nullptr), ba
         if (writeOutputBam) {
             // output file mangement
             std::string writeBamFile = resultPrefix + "." + outputFormat;
-            // open output bam file
+            // Open output BAM file with appropriate mode (bam/cram)
             out = hts_open(writeBamFile.c_str(), (outputFormat == "bam" ? "wb" : "wc" ));
             // load reference file
             hts_set_fai_filename(out, fastaFile.c_str());
@@ -65,6 +80,13 @@ writeOutputBam(writeOutputBam), isReleased(false), in(nullptr), out(nullptr), ba
     }
 }
 
+/**
+ * @brief Template function to check for null pointers
+ * @param ptr Pointer to check
+ * @param errorMessage Error message to display if pointer is null
+ * 
+ * Throws runtime_error if the pointer is null
+ */
 template<typename T>
 void BamFileRAII::checkNullPointer(const T* ptr, const std::string& errorMessage) const {
     if (ptr == nullptr) {
@@ -72,6 +94,12 @@ void BamFileRAII::checkNullPointer(const T* ptr, const std::string& errorMessage
     }
 }
 
+/**
+ * @brief Validates the state of BAM file resources
+ * @return true if all resources are valid, exits program if not
+ * 
+ * Checks that all BAM file pointers are valid and not null
+ */
 bool BamFileRAII::validateState(){
     try{
         checkNullPointer(in, "(validateState) in ptr is nullptr");
@@ -88,6 +116,11 @@ bool BamFileRAII::validateState(){
     }
 }
 
+/**
+ * @brief Writes the current alignment to the output BAM file
+ * 
+ * Writes the alignment record to the output file and handles errors
+ */
 void BamFileRAII::samWriteBam(){
     try{
         int result = sam_write1(out, bamHdr, aln);
@@ -100,12 +133,24 @@ void BamFileRAII::samWriteBam(){
     }
 }
 
+/**
+ * @brief Destructor for BamFileRAII class
+ * 
+ * Ensures proper cleanup of BAM file resources.
+ * Only destroys resources if they haven't been manually released.
+ */
 BamFileRAII::~BamFileRAII(){
     if (!isReleased) {
         destroy();
     }
 }
 
+/**
+ * @brief Destroys all BAM file resources
+ * 
+ * Frees memory and closes file handles in the correct order.
+ * Sets isReleased flag to prevent double cleanup.
+ */
 void BamFileRAII::destroy(){
     if (aln) bam_destroy1(aln);
     if (idx) hts_idx_destroy(idx);
@@ -115,7 +160,13 @@ void BamFileRAII::destroy(){
     isReleased = true;
 }
 
-
+/**
+ * @brief Constructor for HaplotagBamParser class
+ * @param config Configuration parameters for BAM parsing
+ * @param control Control parameters for processing behavior
+ * 
+ * Initializes the BAM parser with configuration and control settings.
+ */
 HaplotagBamParser::HaplotagBamParser(
     const ParsingBamConfig &config,
     const ParsingBamControl &control
@@ -124,21 +175,44 @@ HaplotagBamParser::HaplotagBamParser(
 
 }
 
+/**
+ * @brief Destructor for HaplotagBamParser class
+ * 
+ * Virtual destructor for proper cleanup of derived classes.
+ */
 HaplotagBamParser::~HaplotagBamParser(){
 
 }
 
+/**
+ * @brief Main function for parsing BAM files and performing haplotagging
+ * @param ctx Context containing all necessary data structures and files
+ * 
+ * Orchestrates the entire BAM parsing and haplotagging process.
+ * Validates input parameters, sets up thread pool, and delegates processing
+ * to appropriate methods based on control mode.
+ * 
+ * Processing flow:
+ * 1. Validate input parameters and context
+ * 2. Determine last variant positions for efficient processing
+ * 3. Initialize FASTA parser and thread pool
+ * 4. Process BAM file based on mode (single/multi-thread)
+ * 5. Clean up resources
+ */
 void HaplotagBamParser::parsingBam(BamParserContext& ctx){
     try{
 
+        // Validate that multi-thread mode is not used with output BAM
         if(control.mode == ParsingBamMode::MULTI_THREAD && control.writeOutputBam) {
             throw std::runtime_error("Cannot set Multi Thread mode when writeOutputBam is true");
         }
 
+        // Validate chromosome vector is not empty
         if(ctx.chrVec.size() == 0){
             throw std::runtime_error("chrVec is empty");
         }
 
+        // Validate chromosome length map is not empty
         if(ctx.chrLength.size() == 0){
             throw std::runtime_error("chrLength is empty");
         }
@@ -149,23 +223,24 @@ void HaplotagBamParser::parsingBam(BamParserContext& ctx){
         // reference fasta parser
         FastaParser fastaParser(ctx.fastaFile, ctx.chrVec, last_pos, config.numThreads);
 
-        // init data structure and get core n
+        // Initialize thread pool for parallel processing
         htsThreadPool threadPool = {NULL, 0};
-        // creat thread pool
+        // Create thread pool with specified number of threads
         if (!(threadPool.pool = hts_tpool_init(config.numThreads))) {
             throw std::runtime_error("Error creating thread pool");
         }
 
+        // Process BAM file based on control mode
         switch(control.mode){
             case ParsingBamMode::SINGLE_THREAD:
                 // Process the BAM file using a single thread
-                std::cerr<< "[single thread]";
+                // std::cerr << "[single thread]"; // [debug]
                 processBamWithOutput(ctx, fastaParser, threadPool);
                 break;
 
             case ParsingBamMode::MULTI_THREAD:
                 // Process the BAM file using multiple threads
-                std::cerr<< "[multi thread]";
+                // std::cerr << "[multi thread]"; // [debug]
                 processBamParallel(ctx, fastaParser, threadPool);
                 break;
 
@@ -173,6 +248,7 @@ void HaplotagBamParser::parsingBam(BamParserContext& ctx){
                 throw std::runtime_error("Unsupported parsing bam mode: " + std::to_string(control.mode));
         }
 
+        // Clean up thread pool
         hts_tpool_destroy(threadPool.pool);
     }catch(const std::exception& e){
         std::cerr << "[ERROR](HaplotagBamParser): " << e.what() << std::endl;
@@ -182,6 +258,16 @@ void HaplotagBamParser::parsingBam(BamParserContext& ctx){
     return;
 }
 
+/**
+ * @brief Processes BAM file using parallel processing
+ * @param ctx BAM parser context
+ * @param fastaParser Reference sequence parser
+ * @param threadPool Thread pool for parallel execution
+ * 
+ * Uses OpenMP to process chromosomes in parallel.
+ * Each chromosome is processed independently with its own BAM file resources.
+ * This mode does not support output BAM generation due to thread safety concerns.
+ */
 void HaplotagBamParser::processBamParallel(
     BamParserContext& ctx,
     const FastaParser &fastaParser,
@@ -190,37 +276,63 @@ void HaplotagBamParser::processBamParallel(
     // loop all chromosome
     #pragma omp parallel for schedule(dynamic) num_threads(config.numThreads) 
     for(auto chr : ctx.chrVec ){
-        // bam file resource allocation
+        // Allocate BAM file resources for each thread
         BamFileRAII bamRAII(ctx.BamFile, ctx.fastaFile, threadPool, config.version, config.command, config.resultPrefix, config.outputFormat, control.writeOutputBam);
 
+        // Create chromosome processing context
         ChrProcContext commonCtx(chr, ctx.chrLength.at(chr), config, ctx.genomeSample, ctx.vcfSet);
-        //create the chromosome processor
+        // Create the chromosome processor using factory method
         auto chrProcessor = createProcessor(chr);
-        chrProcessor->processSingleChromosome(commonCtx, bamRAII, fastaParser, ctx.mergedChrVarinat);
+        // Process the chromosome
+        chrProcessor->processSingleChrom(commonCtx, bamRAII, fastaParser, ctx.mergedChrVarinat);
     }
 }
 
+/**
+ * @brief Processes BAM file using single-threaded processing with output
+ * @param ctx BAM parser context
+ * @param fastaParser Reference sequence parser
+ * @param threadPool Thread pool for parallel execution
+ * 
+ * Processes chromosomes sequentially and writes output BAM file.
+ * This mode supports output BAM generation and provides progress information.
+ * Each chromosome is processed one at a time to maintain output order.
+ */
 void HaplotagBamParser::processBamWithOutput(
     BamParserContext& ctx,
     const FastaParser &fastaParser,
     htsThreadPool &threadPool
 ){
-    // bam file resource allocation
+    // Allocate BAM file resources for single-threaded processing
     BamFileRAII bamRAII(ctx.BamFile, ctx.fastaFile, threadPool, config.version, config.command, config.resultPrefix, config.outputFormat, control.writeOutputBam);
     // loop all chromosome
     for(auto chr : ctx.chrVec ){
-        //create the chromosome processor
+        // Create the chromosome processor
         std::time_t begin = time(NULL);
         std::cerr<<"chr: " << chr << " ... " ;
 
+        // Create chromosome processing context
         ChrProcContext commonCtx(chr, ctx.chrLength.at(chr), config, ctx.genomeSample, ctx.vcfSet);
-        //create the chromosome processor
+        // Create the chromosome processor using factory method
         auto chrProcessor = createProcessor(chr);
-        chrProcessor->processSingleChromosome(commonCtx, bamRAII, fastaParser, ctx.mergedChrVarinat);
+        // Process the chromosome
+        chrProcessor->processSingleChrom(commonCtx, bamRAII, fastaParser, ctx.mergedChrVarinat);
         std::cerr<< difftime(time(NULL), begin) << "s\n";
     }
 }
 
+/**
+ * @brief Gets the last variant position for each chromosome
+ * @param last_pos Vector to store last variant positions
+ * @param chrVec Vector of chromosome names
+ * @param mergedChrVarinat Map of variants by chromosome
+ * @param genomeSample Genome type (NORMAL/TUMOR)
+ * 
+ * Determines the last variant position for efficient processing.
+ * For NORMAL genome: finds last phased variant
+ * For TUMOR genome: finds last variant (tumor or phased normal)
+ * If no variants exist, sets position to 0.
+ */
 void HaplotagBamParser::getLastVarPos(
     std::vector<int>& last_pos, 
     const std::vector<std::string>& chrVec, 
@@ -230,8 +342,10 @@ void HaplotagBamParser::getLastVarPos(
     for( auto chr : chrVec ){
         bool existLastPos = false;
 
+        // Search for last variant from end of chromosome
         for (auto lastVariantIter = mergedChrVarinat[chr].rbegin(); lastVariantIter != mergedChrVarinat[chr].rend(); ++lastVariantIter) {
             if (genomeSample == NORMAL) {
+                // For normal genome, find last phased variant
                 if ((*lastVariantIter).second.isExists(NORMAL) && (*lastVariantIter).second.Variant[NORMAL].isExistPhasedSet()) {
                     last_pos.push_back((*lastVariantIter).first);
                         existLastPos = true;
@@ -239,6 +353,7 @@ void HaplotagBamParser::getLastVarPos(
                     }
             }
             else if (genomeSample == TUMOR) {
+                // For tumor genome, find last variant (tumor or phased normal)
                 if ((*lastVariantIter).second.isExists(TUMOR) || 
                     ((*lastVariantIter).second.isExists(NORMAL) && (*lastVariantIter).second.Variant[NORMAL].isExistPhasedSet())) {
                     last_pos.push_back((*lastVariantIter).first);
@@ -246,28 +361,64 @@ void HaplotagBamParser::getLastVarPos(
                     break;
                 }
             }else{
-                std::cerr << "ERROR (getRefLastVarPos) => unsupported genome sample: " << genomeSample << std::endl;
+                std::cerr << "[ERROR] (getRefLastVarPos) => unsupported genome sample: " << genomeSample << std::endl;
                 exit(EXIT_SUCCESS);
             }
         }
         
+        // If no variants found, set position to 0
         if(!existLastPos){
             last_pos.push_back(0);
         }
     }
 }
 
+/**
+ * @brief Constructor for ChromosomeProcessor class
+ * @param writeOutputBam Whether to write output BAM file
+ * @param mappingQualityFilter Whether to filter by mapping quality
+ * 
+ * Initializes chromosome processor with output and filtering settings.
+ */
 ChromosomeProcessor::ChromosomeProcessor(bool writeOutputBam, bool mappingQualityFilter)
 : writeOutputBam(writeOutputBam), mappingQualityFilter(mappingQualityFilter)
 {
 
 }
 
+/**
+ * @brief Destructor for ChromosomeProcessor class
+ * 
+ * Virtual destructor for proper cleanup of derived classes.
+ */
 ChromosomeProcessor::~ChromosomeProcessor(){
 
 }
 
-void ChromosomeProcessor::processSingleChromosome(
+/**
+ * @brief Processes a single chromosome
+ * @param ctx Chromosome processing context
+ * @param bamRAII BAM file RAII wrapper
+ * @param fastaParser Reference sequence parser
+ * @param mergedChrVarinat Map of variants by chromosome
+ * 
+ * Main function for processing all reads in a chromosome.
+ * 
+ * Processing flow:
+ * 1. Extract chromosome variants and set up iterators
+ * 2. Fetch reference sequence for the chromosome
+ * 3. Set up region iterator for chromosome-specific reads
+ * 4. Process each read based on its characteristics:
+ *    - Low mapping quality reads
+ *    - Unmapped reads
+ *    - Secondary alignments
+ *    - Supplementary alignments
+ *    - Reads with no variants
+ *    - Normal reads for processing
+ * 5. Write reads to output BAM if enabled
+ * 6. Perform post-processing operations
+ */
+void ChromosomeProcessor::processSingleChrom(
     ChrProcContext& ctx,
     BamFileRAII& bamRAII,
     const FastaParser& fastaParser,
@@ -277,7 +428,7 @@ void ChromosomeProcessor::processSingleChromosome(
     const int& chrLength = ctx.chrLength;
     const ParsingBamConfig& params = ctx.params; 
 
-    // records all variants within this chromosome.
+    // Extract all variants within this chromosome for processing
     std::map<int, MultiGenomeVar> currentVariants;
 
     #pragma omp critical
@@ -294,24 +445,26 @@ void ChromosomeProcessor::processSingleChromosome(
     // fetch chromosome string
     std::string ref_string = fastaParser.chrString.at(chr);
 
+    // Set up region for chromosome-specific read iteration
     std::string region = !params.region.empty() ? params.region : std::string(chr + ":1-" + std::to_string(chrLength));
     hts_itr_t* iter = sam_itr_querys(bamRAII.idx, bamRAII.bamHdr, region.c_str());
 
+    // Process each read in the chromosome
     while (sam_itr_multi_next(bamRAII.in, iter, bamRAII.aln) >= 0) {
         
         int flag = bamRAII.aln->core.flag;
 
         if ( bamRAII.aln->core.qual < params.qualityThreshold && mappingQualityFilter){
-           // mapping quality is lower than threshold
+           // Mapping quality is lower than threshold
            processLowMappingQuality();
         }
         else if( (flag & 0x4) != 0 ){
-            // read unmapped
+            // Read is unmapped
             processUnmappedRead();
         }
         else if( (flag & 0x100) != 0 ){
-            // secondary alignment. repeat.
-            // A secondary alignment occurs when a given read could align reasonably well to more than one place.
+            // Secondary alignment
+            // A secondary alignment occurs when a given read could align reasonably well to more than one place
             processSecondaryAlignment();
         }
         else if( (flag & 0x800) != 0 && params.tagSupplementary == false ){
@@ -319,7 +472,7 @@ void ChromosomeProcessor::processSingleChromosome(
             // A chimeric alignment is represented as a set of linear alignments that do not have large overlaps.
             processSupplementaryAlignment();
         }
-        //currentVariants rbegin == rend
+        // Check if no variants exist for this chromosome
         else if(last == currentVariants.rend()){ 
             //skip this read
             processEmptyVariants();
@@ -331,77 +484,59 @@ void ChromosomeProcessor::processSingleChromosome(
             processOtherCase();
         }
 
-        //for write all reads in bam file
+        // Write all reads to output BAM file if enabled
         if(writeOutputBam){
             bamRAII.samWriteBam();
         }
     }
 
+    // Perform post-processing operations for the chromosome
     postProcess(chr, currentVariants);
     
+    // Clean up region iterator
     hts_itr_destroy(iter);
 }
 
-void ChromosomeProcessor::calculateBaseCommonInfo(PosBase& baseInfo, std::string& tumorAltBase){
-    int &depth = baseInfo.depth;
-    int AltCount = baseInfo.getBaseCount(tumorAltBase);
-
-    int filteredMpqDepth = baseInfo.filteredMpqDepth;
-    int filteredMpqAltCount = baseInfo.getMpqBaseCount(tumorAltBase);
-
-    baseInfo.VAF = calculateVAF(AltCount, depth);
-    baseInfo.filteredMpqVAF = calculateVAF(filteredMpqAltCount, filteredMpqDepth);
-    baseInfo.nonDelVAF = calculateVAF(AltCount, (depth - baseInfo.delCount));
-
-    baseInfo.lowMpqReadRatio = calculateLowMpqReadRatio(depth, filteredMpqDepth);
-    baseInfo.delRatio = calculateDelRatio(baseInfo.delCount, depth);
-
-    //read hp count in the normal bam
-    int H1readCount = baseInfo.ReadHpCount[ReadHP::H1];
-    int H2readCount = baseInfo.ReadHpCount[ReadHP::H2];
-    int germlineReadHpCount = H1readCount + H2readCount;
-
-    baseInfo.germlineHaplotypeImbalanceRatio = calculateHaplotypeImbalanceRatio(H1readCount, H2readCount, germlineReadHpCount);
-
-    baseInfo.percentageOfGermlineHp = calculatePercentageOfGermlineHp(germlineReadHpCount, depth);
-}
-
-float ChromosomeProcessor::calculateVAF(int altCount, int depth){
-    return (depth == 0 || altCount == 0) ? 0.0 : (float)altCount / (float)depth;
-}
-
-float ChromosomeProcessor::calculateLowMpqReadRatio(int depth, int filteredMpqDepth){
-    return depth == 0 ? 0.0 : (float)(depth - filteredMpqDepth) / (float)depth;
-}
-
-float ChromosomeProcessor::calculateDelRatio(int delCount, int depth){
-    return (depth == 0 || delCount == 0) ? 0.0 : (float)delCount / (float)depth;
-}
-
-double ChromosomeProcessor::calculateHaplotypeImbalanceRatio(int& H1readCount, int& H2readCount, int& totalReadCount) {
-    if(H1readCount > 0 && H2readCount > 0) {
-        return (H1readCount > H2readCount) ? 
-               ((double)H1readCount / (double)totalReadCount) : 
-               ((double)H2readCount / (double)totalReadCount);
-    } else if(H1readCount == 0 && H2readCount == 0) {
-        return 0.0;
-    } else {
-        return 1.0;
-    }
-}
-
-double ChromosomeProcessor::calculatePercentageOfGermlineHp(int& totalGermlineReadCount, int& depth) {
-    return (depth == 0 || totalGermlineReadCount == 0) ? 0.0 : (double)totalGermlineReadCount / (double)depth;
-}
-
+/**
+ * @brief Constructor for CigarParser class
+ * @param ctx CIGAR parsing context
+ * @param ref_pos Reference position (will be modified)
+ * @param query_pos Query position (will be modified)
+ * 
+ * Initializes CIGAR parser with context and position references.
+ */
 CigarParser::CigarParser(CigarParserContext& ctx, int& ref_pos, int& query_pos)
 : ctx(ctx), ref_pos(ref_pos), query_pos(query_pos){}
 
+/**
+ * @brief Destructor for CigarParser class
+ * 
+ * Virtual destructor for proper cleanup of derived classes.
+ */
 CigarParser::~CigarParser(){
 
 };
 
-
+/**
+ * @brief Parses CIGAR string and processes alignment operations
+ * @param hpCount Map to count haplotype assignments
+ * @param variantsHP Map to record variant haplotype assignments
+ * @param norCountPS Map to count phase set assignments
+ * 
+ * Main function for parsing CIGAR operations and determining haplotype assignments.
+ * 
+ * CIGAR parsing flow:
+ * 1. Skip variants that are to the left of the read start position
+ * 2. Initialize position trackers (reference and query positions)
+ * 3. Iterate through each CIGAR operation:
+ *    - Match operations (M, =, X): process variants within match region
+ *    - Insertion operations (I): skip reference position, advance query position
+ *    - Deletion operations (D): process variants within deletion region
+ *    - Skipped operations (N): advance reference position only
+ *    - Soft clipping (S): advance query position only
+ *    - Hard clipping (H): no position advancement
+ *    - Padding (P): no position advancement
+ */
 void CigarParser::parsingCigar(
     std::map<int, int>& hpCount,
     std::map<int, int>& variantsHP,
@@ -427,7 +562,7 @@ void CigarParser::parsingCigar(
     // position relative to read
     query_pos = 0;
 
-    // reading cigar to detect snp on this read
+    // Parse CIGAR string to detect variants on this read
     int aln_core_n_cigar = int(ctx.aln.core.n_cigar);
     for (int i = 0; i < aln_core_n_cigar; i++) {
         uint32_t* cigar = bam_get_cigar(&ctx.aln);
@@ -496,6 +631,15 @@ void CigarParser::parsingCigar(
     }
 }
 
+/**
+ * @brief Counts base nucleotides and applies mapping quality filtering
+ * @param posBase Position base information structure
+ * @param base Base nucleotide
+ * @param aln BAM alignment record
+ * @param mpqThreshold Mapping quality threshold
+ * 
+ * Updates base counts and filtered depth based on mapping quality
+ */
 void CigarParser::countBaseNucleotide(PosBase& posBase, std::string& base, const bam1_t& aln, const float& mpqThreshold){
     // mapping quality is higher than threshold
     if ( aln.core.qual >= mpqThreshold ){
@@ -526,549 +670,13 @@ void CigarParser::countBaseNucleotide(PosBase& posBase, std::string& base, const
     posBase.depth++;  
 }
 
+/**
+ * @brief Counts deletion bases
+ * @param posBase Position base information structure
+ * 
+ * Updates deletion count and total depth
+ */
 void CigarParser::countDeletionBase(PosBase& posBase){
     posBase.delCount++;
     posBase.depth++;
-}
-
-
-
-void GermlineHaplotagStrategy::judgeSnpHap(
-    const std::string& chrName,
-    VarData& norVar,
-    const std::string& base,
-    int& ref_pos,
-    int& length,
-    int& i,
-    int& aln_core_n_cigar,
-    uint32_t* cigar,
-    std::map<int, MultiGenomeVar>::iterator &currentVariantIter,
-    std::map<int, int>& hpCount,
-    std::map<int, int>& variantsHP,
-    std::map<int, int>& countPS
-){
-    int curPos = (*currentVariantIter).first;
-
-    // currentVariant is SNP
-    if( norVar.variantType == VariantType::SNP ){
-        // Detected that the base of the read is either REF or ALT. 
-        if( (base == norVar.allele.Ref) || (base == norVar.allele.Alt) ){
-
-
-            if(!norVar.isExistPhasedSet()){
-                std::cerr << "ERROR (judgeSnpHap) => can't find the position:" 
-                          << " chr: " << chrName << "\t"
-                          << " pos: " << curPos << "\t"
-                          << " ref: " << norVar.allele.Ref << "\t"
-                          << " alt: " << norVar.allele.Alt << "\n";
-                exit(EXIT_SUCCESS);
-            }
-            else{
-                if( base == norVar.HP1){
-                    hpCount[SnpHP::GERMLINE_H1]++;
-                    variantsHP[curPos]=0;
-                }
-                if( base == norVar.HP2){
-                    hpCount[SnpHP::GERMLINE_H2]++;
-                    variantsHP[curPos]=1;
-                }
-                countPS[norVar.PhasedSet]++;
-            }
-            
-        }
-    }
-    // currentVariant is insertion
-    else if( norVar.variantType == VariantType::INSERTION && i+1 < aln_core_n_cigar){
-        
-        int hp1Length = norVar.HP1.length();
-        int hp2Length = norVar.HP2.length();
-        
-        if ( ref_pos + length - 1 == (*currentVariantIter).first && bam_cigar_op(cigar[i+1]) == 1 ) {
-            // hp1 occur insertion
-            if( hp1Length != 1 && hp2Length == 1 ){
-                hpCount[SnpHP::GERMLINE_H1]++;
-                variantsHP[curPos]=0;
-            }
-            // hp2 occur insertion
-            else if( hp1Length == 1 && hp2Length != 1 ){
-                hpCount[SnpHP::GERMLINE_H2]++;
-                variantsHP[curPos]=1;
-            }
-        }
-        else {
-            // hp1 occur insertion
-            if( hp1Length != 1 && hp2Length == 1 ){
-                hpCount[SnpHP::GERMLINE_H2]++;
-                variantsHP[curPos]=1;
-            }
-            // hp2 occur insertion
-            else if( hp1Length == 1 && hp2Length != 1 ){
-                hpCount[SnpHP::GERMLINE_H1]++;
-                variantsHP[curPos]=0;
-            }
-        }
-        countPS[norVar.PhasedSet]++;
-    } 
-    // currentVariant is deletion
-    else if( norVar.variantType == VariantType::DELETION && i+1 < aln_core_n_cigar) {
-
-        int hp1Length = norVar.HP1.length();
-        int hp2Length = norVar.HP2.length();
-        
-        if ( ref_pos + length - 1 == (*currentVariantIter).first && bam_cigar_op(cigar[i+1]) == 2 ) {
-            // hp1 occur deletion
-            if( hp1Length != 1 && hp2Length == 1 ){
-                hpCount[SnpHP::GERMLINE_H1]++;
-                variantsHP[curPos]=0;
-            }
-            // hp2 occur deletion
-            else if( hp1Length == 1 && hp2Length != 1 ){
-                hpCount[SnpHP::GERMLINE_H2]++;
-                variantsHP[curPos]=1;
-            }
-        }
-        else {
-            // hp2 occur deletion
-            if( hp1Length != 1 && hp2Length == 1 ){
-                hpCount[SnpHP::GERMLINE_H2]++;
-                variantsHP[curPos]=1;
-            }
-            // hp1 occur deletion
-            else if( hp1Length == 1 && hp2Length != 1 ){
-                hpCount[SnpHP::GERMLINE_H1]++;
-                variantsHP[curPos]=0;
-            }
-        }
-        countPS[norVar.PhasedSet]++;
-    } 
-}
-
-
-void GermlineHaplotagStrategy::judgeDeletionHap(
-    const std::string& chrName,
-    const std::string& ref_string,
-    int& ref_pos,
-    int& length,
-    int& query_pos,
-    std::map<int, MultiGenomeVar>::iterator &currentVariantIter,
-    const bam1_t* aln,
-    std::map<int, int>& hpCount,
-    std::map<int, int>& variantsHP,
-    std::map<int, int>& countPS
-) {
-    if (ref_string != "") {
-        int del_len = length;
-        if (ref_pos + del_len + 1 == (*currentVariantIter).first) {
-            //if (homopolymerLength((*currentVariantIter).first, ref_string) >= 3) {
-                // special case
-            //}
-        } else if ((*currentVariantIter).first >= ref_pos && (*currentVariantIter).first < ref_pos + del_len) {
-            // check variant in homopolymer
-            if (homopolymerLength((*currentVariantIter).first, ref_string) >= 3) {
-                
-                int curPos = (*currentVariantIter).first;
-                auto norVar = (*currentVariantIter).second.Variant[NORMAL];
-                
-                // SNP
-                if (norVar.variantType == VariantType::SNP) {
-                    // get the next match
-                    char base_chr = seq_nt16_str[bam_seqi(bam_get_seq(aln), query_pos)];
-                    std::string base(1, base_chr);
-
-                    if (base == norVar.HP1) {
-                        hpCount[SnpHP::GERMLINE_H1]++;
-                        variantsHP[curPos] = 0;
-                    }
-                    if (base == norVar.HP2) {
-                        hpCount[SnpHP::GERMLINE_H2]++;
-                        variantsHP[curPos] = 1;
-                    }
-                    countPS[norVar.PhasedSet]++;
-                }
-                
-                // the read deletion contain VCF's deletion
-                else if (norVar.variantType == VariantType::DELETION) {
-
-                    int hp1Length = norVar.HP1.length();
-                    int hp2Length = norVar.HP2.length();
-                    // hp1 occur deletion
-                    if (hp1Length != 1 && hp2Length == 1) {
-                        hpCount[SnpHP::GERMLINE_H1]++;
-                        variantsHP[curPos] = 0;
-                    }
-                    // hp2 occur deletion
-                    else if (hp1Length == 1 && hp2Length != 1) {
-                        hpCount[SnpHP::GERMLINE_H2]++;
-                        variantsHP[curPos] = 1;
-                    }
-                    countPS[norVar.PhasedSet]++;
-                }
-            }
-        }
-    }
-}
-
-void GermlineHaplotagStrategy::judgeSVHap(const bam1_t &aln, std::map<Genome, VCF_Info> &vcfSet, std::map<int, int>& hpCount, const int& genomeSample){
-    auto readIter = vcfSet[Genome::NORMAL].readSVHapCount.find(bam_get_qname(&aln));
-    if( readIter != vcfSet[Genome::NORMAL].readSVHapCount.end()){
-        hpCount[SnpHP::GERMLINE_H1] += vcfSet[Genome::NORMAL].readSVHapCount[bam_get_qname(&aln)][0];
-        hpCount[SnpHP::GERMLINE_H2] += vcfSet[Genome::NORMAL].readSVHapCount[bam_get_qname(&aln)][1];
-    }
-}
-
-int GermlineHaplotagStrategy::judgeReadHap(
-    std::map<int, int>& hpCount, 
-    double& min, 
-    double& max, 
-    double& percentageThreshold, 
-    int& pqValue, 
-    int& psValue, 
-    std::map<int, int>& countPS, 
-    int* totalHighSimilarity, 
-    int* totalWithOutVaraint
-){
-    int hpResult = ReadHP::unTag;
-
-    if(hpCount[SnpHP::GERMLINE_H1] > hpCount[SnpHP::GERMLINE_H2]){
-        min = hpCount[SnpHP::GERMLINE_H2];
-        max = hpCount[SnpHP::GERMLINE_H1];
-    }
-    else{
-        min = hpCount[SnpHP::GERMLINE_H1];
-        max = hpCount[SnpHP::GERMLINE_H2];
-    }
-
-    if( max/(max+min) < percentageThreshold){
-        // no tag
-        pqValue = 0;
-        if(totalHighSimilarity != nullptr) (*totalHighSimilarity)++;
-    }
-    else{
-        if(hpCount[SnpHP::GERMLINE_H1] > hpCount[SnpHP::GERMLINE_H2]){
-            hpResult = ReadHP::H1;
-        }
-        if(hpCount[SnpHP::GERMLINE_H1] < hpCount[SnpHP::GERMLINE_H2]){
-            hpResult = ReadHP::H2;
-        }
-    }
-
-    if( max == 0 ){
-        pqValue=0;
-        if(totalWithOutVaraint != nullptr) (*totalWithOutVaraint)++;
-    }
-    else if( max == ( max + min ) ){
-        pqValue=40;
-    }
-    else{
-        pqValue=-10*(std::log10((double)min/double(max+min)));
-    }
-    
-    // cross two block
-    if( countPS.size() > 1  ){
-        hpResult = ReadHP::unTag;
-    }
-    //set psValue
-    if(hpResult != ReadHP::unTag){
-        auto psIter = countPS.begin();
-        psValue = (*psIter).first;
-    }
-    return hpResult;
-}
-
-void SomaticJudgeHapStrategy::judgeSomaticSnpHap(std::map<int, MultiGenomeVar>::iterator &currentVariantIter, std::string chrName, std::string base, std::map<int, int> &hpCount, std::map<int, int> &norCountPS, std::map<int, int> &tumCountPS, std::map<int, int> *variantsHP, std::vector<int> *tumorAllelePosVec){
-    int curPos = (*currentVariantIter).first;
-    auto& curVar = (*currentVariantIter).second;
-
-    // normal & tumor SNP at the current position (base on normal phased SNPs)
-    // both normal and tumor samples that do not exist in the high-confidence set
-    if(curVar.isExists(NORMAL) && curVar.isExists(TUMOR)){
-        // the tumor & normal SNP GT are phased heterozygous 
-        if((curVar.Variant[NORMAL].GT == GenomeType::PHASED_HETERO) && 
-            (curVar.Variant[TUMOR].GT == GenomeType::PHASED_HETERO)){ 
-            judgeNormalSnpHap(chrName, curPos, curVar, base, hpCount, norCountPS, variantsHP);
-
-        }
-        //the normal SNP GT is phased heterozgous & the tumor SNP GT is unphased heterozgous 
-        else if((curVar.Variant[NORMAL].GT == GenomeType::PHASED_HETERO) && 
-                (curVar.Variant[TUMOR].GT == GenomeType::UNPHASED_HETERO)){   
-            judgeNormalSnpHap(chrName, curPos, curVar, base, hpCount, norCountPS, variantsHP);
-
-        }
-        //the normal SNP GT is phased heterozgous & the tumor SNP GT is homozygous 
-        else if((curVar.Variant[NORMAL].GT == GenomeType::PHASED_HETERO) && 
-                (curVar.Variant[TUMOR].GT == GenomeType::UNPHASED_HOMO)){ 
-            judgeNormalSnpHap(chrName, curPos, curVar, base, hpCount, norCountPS, variantsHP);
-        }
-    // only normal SNP at the current position
-    }else if(curVar.isExists(NORMAL)){
-        // the normal SNP GT is phased heterozgous SNP
-        if((curVar.Variant[NORMAL].GT == GenomeType::PHASED_HETERO)){
-            judgeNormalSnpHap(chrName, curPos, curVar, base, hpCount, norCountPS, variantsHP);
-        }
-    // only tumor SNP at the current position
-    }else if(curVar.isExists(TUMOR)){
-        //the tumor SNP GT is phased heterozygous
-        if(curVar.Variant[TUMOR].GT == GenomeType::PHASED_HETERO){
-            if(curVar.Variant[TUMOR].allele.Ref == base || curVar.Variant[TUMOR].allele.Alt == base){
-                if(!curVar.Variant[TUMOR].isExistPhasedSet()){
-                    std::cerr<< curPos << "\t"
-                             << curVar.Variant[TUMOR].allele.Ref << "\t"
-                             << curVar.Variant[TUMOR].allele.Alt << "\n";
-                    exit(EXIT_SUCCESS);
-                }else{
-                    judgeTumorOnlySnpHap(chrName, curPos, curVar, base, hpCount, &tumCountPS, variantsHP, tumorAllelePosVec);
-                }
-            }
-        //the tumor SNP GT is unphased heterozygous
-        }else if(curVar.Variant[TUMOR].GT == GenomeType::UNPHASED_HETERO){
-            if(curVar.Variant[TUMOR].allele.Ref == base || curVar.Variant[TUMOR].allele.Alt == base){
-                judgeTumorOnlySnpHap(chrName, curPos, curVar, base, hpCount, nullptr, variantsHP, tumorAllelePosVec);
-            }           
-        //the tumor SNP GT is homozygous
-        }else if(curVar.Variant[TUMOR].GT == GenomeType::UNPHASED_HOMO){
-            if(curVar.Variant[TUMOR].allele.Ref == base || curVar.Variant[TUMOR].allele.Alt == base){
-                judgeTumorOnlySnpHap(chrName, curPos, curVar, base, hpCount, nullptr, variantsHP, tumorAllelePosVec);
-            }
-        }
-    }
-}
-
-void SomaticJudgeHapStrategy::judgeNormalSnpHap(
-    const std::string& chrName, 
-    int& curPos,
-    MultiGenomeVar& curVar,
-    std::string& base,
-    std::map<int, int>& hpCount, 
-    std::map<int, int>& norCountPS,
-    std::map<int, int> *variantsHP
-){
-    if(curVar.Variant[NORMAL].allele.Ref == base || curVar.Variant[NORMAL].allele.Alt == base){
-
-        if(!curVar.Variant[NORMAL].isExistPhasedSet()){
-            std::cerr<< "Unable to locate the phase set of the current normal SNP\n"
-                        << curPos << "\t"
-                        << curVar.Variant[NORMAL].allele.Ref << "\t"
-                        << curVar.Variant[NORMAL].allele.Alt  << "\n";
-            exit(EXIT_SUCCESS);
-        }
-
-        std::string& norHP1 = curVar.Variant[NORMAL].HP1;
-        std::string& norHP2 = curVar.Variant[NORMAL].HP2;
-
-        if( base == norHP1){
-            hpCount[1]++;
-            if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::GERMLINE_H1;
-        }
-        if(base == norHP2){
-            hpCount[2]++;
-            if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::GERMLINE_H2;
-        }
-        norCountPS[curVar.Variant[NORMAL].PhasedSet]++;
-    }
-}
-
-int SomaticJudgeHapStrategy::judgeSomaticReadHap(
-    std::map<int, int> &hpCount,
-    int &pqValue,
-    std::map<int, int> &norCountPS,
-    double &norHPsimilarity,
-    double &tumHPsimilarity,
-    double percentageThreshold,
-    int *totalHighSimilarity,
-    int *totalCrossTwoBlock,
-    int *totalWithOutVaraint
-){
-    double normalMinHPcount = 0;
-    double normalMaxHPcount = 0;
-    int maxNormalHP = 0;
-
-    double tumorMinHPcount = 0;
-    double tumorMaxHPcount = 0;
-    int maxTumorHP = 0;
-
-    // determine max and min
-    //tumor HP count
-    if(hpCount[3] > hpCount[4]){
-        tumorMinHPcount = hpCount[4];
-        tumorMaxHPcount = hpCount[3];
-        maxTumorHP = SnpHP::SOMATIC_H3;
-    }
-    else{
-        tumorMinHPcount = hpCount[3];
-        tumorMaxHPcount = hpCount[4];
-        maxTumorHP = SnpHP::SOMATIC_H4;
-    }
-
-    //normal HP count
-    if(hpCount[1] > hpCount[2]){
-        normalMinHPcount = hpCount[2];
-        normalMaxHPcount = hpCount[1];
-        maxNormalHP = SnpHP::GERMLINE_H1;
-    }
-    else{
-        normalMinHPcount = hpCount[1];
-        normalMaxHPcount = hpCount[2];
-        maxNormalHP = SnpHP::GERMLINE_H2;
-    }
-
-    //the similarity of HP types
-    //tumor variants
-    tumHPsimilarity = (tumorMaxHPcount == 0) ? 0.0 : tumorMaxHPcount/(tumorMaxHPcount+tumorMinHPcount);
-    norHPsimilarity = (normalMaxHPcount == 0) ? 0.0 : normalMaxHPcount/(normalMaxHPcount+normalMinHPcount);
-
-
-    // determine the haplotype of the read
-    int hpResult = ReadHP::unTag;
-
-    // read have tumor variant
-    if(tumorMaxHPcount != 0){
-        //check the similarity of HP types in the tumor
-        if(tumHPsimilarity >= percentageThreshold){
-            //check the similarity of HP types in the normal
-            if(norHPsimilarity >= percentageThreshold){
-                switch(maxTumorHP){
-                    case SnpHP::SOMATIC_H3:
-                        if(maxNormalHP == SnpHP::GERMLINE_H1){
-                            hpResult = ReadHP::H1_1;
-                        }else if(maxNormalHP == SnpHP::GERMLINE_H2){
-                            hpResult = ReadHP::H2_1;
-                        }
-                        break;
-                    case SnpHP::SOMATIC_H4:
-                        if(maxNormalHP == SnpHP::GERMLINE_H1){
-                            hpResult = ReadHP::H1_2;
-                        }else if(maxNormalHP == SnpHP::GERMLINE_H2){
-                            hpResult = ReadHP::H2_2;
-                        }
-                        break;
-                    default:
-                        std::cerr << "ERROR: Unexpected haplotype : tumor Max HP= "<< maxTumorHP << std::endl;
-                        exit(1);
-                        break;
-                }
-            }
-            // can't determine HP from germline
-            else{
-                //The read has a somatic SNP, but the germline haplotype(H1/H2) it came from is unknown
-                switch(maxTumorHP){
-                    case SnpHP::SOMATIC_H3:
-                        hpResult = ReadHP::H3; break;
-                    case SnpHP::SOMATIC_H4:
-                        hpResult = ReadHP::H4; break;
-                    default:
-                        std::cerr << "ERROR: Unexpected haplotype : tumor Max HP= "<< maxTumorHP << std::endl;
-                        exit(1);
-                        break;
-                }
-            }
-        }else{
-            // no tag
-            pqValue = 0;
-            if(totalHighSimilarity != nullptr) (*totalHighSimilarity)++;
-        }
-    }
-    // read haven't tumor variant
-    else if(normalMaxHPcount != 0){
-        if(norHPsimilarity >= percentageThreshold){
-            hpResult = maxNormalHP;
-        }else{
-            // no tag
-            pqValue = 0;
-            if(totalHighSimilarity != nullptr) (*totalHighSimilarity)++;
-        }
-    }
-
-    // cross two block
-    // had at least one tumor-unique SNP in the current read
-    if(hpCount[3] != 0){
-        if(norCountPS.size() > 1){
-            hpResult = ReadHP::unTag;
-            if(totalCrossTwoBlock != nullptr) (*totalCrossTwoBlock)++;
-        }
-    // all SNPs are germline variants in the current read
-    }else{
-        if(norCountPS.size() > 1){
-            hpResult = ReadHP::unTag;
-            if(totalCrossTwoBlock != nullptr) (*totalCrossTwoBlock)++;
-        }
-    }
-
-    // determine the quality of the read
-    // There haven't been any variants in the current read
-    if(normalMaxHPcount == 0 && tumorMaxHPcount == 0){
-        //std::cerr<< "Read max = 0 : " << bam_get_qname(&aln) << "\n";
-        if(totalWithOutVaraint != nullptr) (*totalWithOutVaraint)++;
-        pqValue=0;
-    }
-    // read have tumor variant
-    else if(tumorMaxHPcount != 0){
-        if( tumorMaxHPcount == ( tumorMaxHPcount + tumorMinHPcount ) ){
-            pqValue=40;
-        }
-        else{
-            pqValue=-10*(std::log10((double)tumorMinHPcount/double(tumorMaxHPcount+tumorMinHPcount)));
-        }
-    }
-    else if(normalMaxHPcount != 0){
-        if( normalMaxHPcount == ( normalMaxHPcount + normalMinHPcount ) ){
-            pqValue=40;
-        }
-        else{
-            pqValue=-10*(std::log10((double)normalMinHPcount/double(normalMaxHPcount+normalMinHPcount)));
-        }
-    }
-
-    return hpResult;
-}
-
-
-void ExtractSomaticDataStragtegy::judgeTumorOnlySnpHap(const std::string &chrName, int &curPos, MultiGenomeVar &curVar, std::string base, std::map<int, int> &hpCount, std::map<int, int> *tumCountPS, std::map<int, int> *variantsHP, std::vector<int> *tumorAllelePosVec){
- //the tumor SNP GT is phased heterozygous
-    //all bases of the same type at the current position in normal.bam
-    if(tumorAllelePosVec == nullptr){
-        std::cerr << "ERROR (SomaticDetectJudgeHP) => tumorAllelePosVec pointer cannot be nullptr"<< std::endl;
-        exit(1);
-    }
-
-    std::string& TumorRefBase = curVar.Variant[TUMOR].allele.Ref;
-    std::string& TumorAltBase = curVar.Variant[TUMOR].allele.Alt; 
-    
-    if(base == TumorAltBase){
-        hpCount[3]++;
-        if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::SOMATIC_H3;
-
-        //record postions that tagged as HP3 for calculating the confidence of somatic positions
-        (*tumorAllelePosVec).push_back(curPos);
-
-    //base is not match to TumorRefBase & TumorAltBase (other HP)
-    }else if(base != TumorRefBase && base != TumorAltBase){
-        //hpCount[4]++;
-        //if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::SOMATIC_H4;
-        //(*tumorAllelePosVec).push_back(curPos);
-    }
-
-    if(tumCountPS != nullptr) (*tumCountPS)[curVar.Variant[TUMOR].PhasedSet]++;
-}
-
-void SomaticHaplotagStrategy::judgeTumorOnlySnpHap(const std::string &chrName, int &curPos, MultiGenomeVar &curVar, std::string base, std::map<int, int> &hpCount, std::map<int, int> *tumCountPS, std::map<int, int> *variantsHP, std::vector<int> *tumorAllelePosVec){
-
-    if(curVar.isSomaticVariant){
-        std::string& TumorRefBase = curVar.Variant[TUMOR].allele.Ref;
-        std::string& TumorAltBase = curVar.Variant[TUMOR].allele.Alt; 
-
-        if(base == TumorAltBase){
-            hpCount[3]++;
-            if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::SOMATIC_H3;
-
-        //base is not match to TumorRefBase & TumorAltBase (other HP)
-        }else if(base != TumorRefBase && base != TumorAltBase){
-            //hpCount[4]++;
-            //if(variantsHP != nullptr) (*variantsHP)[curPos] = SnpHP::SOMATIC_H4;
-            //std::cerr << "Somatic SNP: " << curPos << " " << base << " -> " << TumorRefBase << "|" << TumorAltBase << std::endl;
-        }
-
-        if(curVar.Variant[TUMOR].GT == GenomeType::PHASED_HETERO){
-            if(tumCountPS != nullptr) (*tumCountPS)[curVar[TUMOR].PhasedSet]++;
-        }
-    }
 }
